Se ajusta la interfaz ClientesDao el parametro de tipo Integer al tipo Long,el tipo integer generaba un error al tratar de eliminar un cliente de la base de datos

+++ TiendaE-comerce/src/main/java/com/mintic/TiendaEcomerce/api/ClientesAPI.java
@@ -37,7 +37,7 @@ public class ClientesAPI
 	}
 	
 	@DeleteMapping("/eliminar/{id}")
-	public void eliminar(@PathVariable("id")Integer id)
+	public void eliminar(@PathVariable("id")Long id)
 	{
 		clientesDao.deleteById(id);
 	}
